requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
structlog==23.2.0
sentence-transformers>=2.7.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.4
aiofiles>=23.1.0
httpx==0.25.2
twilio==8.10.3
python-telegram-bot==20.7
openai==1.3.8
groq==0.4.1
google-generativeai==0.3.2
PyPDF2==3.0.1
python-magic==0.4.27
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0